(move) amount component
